Changed build system to modified Slashdev
The basic build system involved is described in fair detail at:

http://www.slashdev.ca/2007/05/30/blackberry-development-with-ant-eclipse/

This system is usable in our (DXY Solutions) modified Slashdev build
environment, which includes defining which versions of the BlackBerry
JDE are supported (via the jde.supports property in project.properties),
and is also usable in an unmodified Slashdev-based build.

Most of our projects are developed on non-Windows hosts, and the
flexibility offered by bb-ant-tools is such that we cannot pass up using
it as our primary system.